Move HistoricalTx to its own file
