+++ scripts/wallet.js
@@ -3,7 +3,8 @@ import { decrypt } from './aes-gcm.js';
 import { beforeUnloadListener } from './global.js';
 import { getNetwork } from './network.js';
 import { MAX_ACCOUNT_GAP } from './chain_params.js';
-import { HistoricalTx, HistoricalTxType, Mempool } from './mempool.js';
+import { Mempool } from './mempool.js';
+import { HistoricalTx, HistoricalTxType } from './historical_tx.js';
 import { COutpoint, Transaction } from './transaction.js';
 import { confirmPopup, createAlert } from './misc.js';
 import { cChainParams } from './chain_params.js';
